fix(core): unify shared space fetchers and coherence priorities
Move space detail web3 helpers into a shared module for both client/server usage
and centralize coherence priorities in ui-utils to prevent drift across packages.
