+++ packages/core/src/space/client/web3/dao-space-factory/get-space-details.ts
@@ -1,21 +1,1 @@
-import {
-  daoSpaceFactoryImplementationAbi,
-  daoSpaceFactoryImplementationAddress,
-} from '@hypha-platform/core/generated';
-
-export const getSpaceDetails = ({
-  spaceId,
-  chain = 8453,
-}: {
-  spaceId: bigint;
-  chain?: keyof typeof daoSpaceFactoryImplementationAddress;
-}) => {
-  const address = daoSpaceFactoryImplementationAddress[chain];
-
-  return {
-    address,
-    abi: daoSpaceFactoryImplementationAbi,
-    functionName: 'getSpaceDetails',
-    args: [spaceId],
-  } as const;
-};
+export { getSpaceDetails } from '../../../shared/web3/get-space-details';
