+++ packages/core/src/space/server/web3/fetch-space-details.ts
@@ -1,3 +1,3 @@
 import 'server-only';
 
-export { fetchSpaceDetails } from '../../client/web3/fetch/fetchSpaceDetails';
+export { fetchSpaceDetails } from '../../shared/web3/fetch-space-details';
