+++ packages/storage-postgres/package.json
@@ -14,6 +14,7 @@
     "vitest": "^1.3.1"
   },
   "dependencies": {
+    "@hypha-platform/ui-utils": "workspace:*",
     "drizzle-kit": "^0.30.5",
     "drizzle-orm": "^0.38.4",
     "drizzle-seed": "^0.3.1",
