+++ packages/ui-utils/src/coherence-priorities.ts
@@ -1,0 +1,8 @@
+export const COHERENCE_PRIORITIES = [
+  'critical',
+  'high',
+  'medium',
+  'low',
+] as const;
+
+export type CoherencePriority = (typeof COHERENCE_PRIORITIES)[number];
