+++ packages/ui-utils/src/index.ts
@@ -15,3 +15,4 @@ export * from './canConvertToBigInt';
 export * from './formatSeconds';
 export * from './formatDecayInterval';
 export * from './stripDescription';
+export * from './coherence-priorities';
